+++ DataLayer/AdoNet/SecurityGroupDao.vb
@@ -108,8 +108,8 @@ Namespace AdoNet
                                     Function(reader) _
             New GroupAccess() With {
             .IdNo = Extensions.AsId(Of Int32)(reader("IdNo")),
-            .SecurityGroupIdNo = Extensions.AsInt(Of Integer?)(reader("SecurityGroupIdNo")),
-            .SecurityObjectIdNo = Extensions.AsInt(Of Integer?)(reader("SecurityObjectIdNo")),
+            .SecurityGroupIdNo = Extensions.AsInt(Of Int16?)(reader("SecurityGroupIdNo")),
+            .SecurityObjectIdNo = Extensions.AsInt(Of Int16?)(reader("SecurityObjectIdNo")),
             .SecurityObjectName = Extensions.AsString(reader("SecurityObjectName")),
             .Visible = Extensions.AsBool(reader("Visible")),
             .Editable = Extensions.AsBool(reader("Editable"))
